better separation of concerns in fixtures and tests
